Missing sys.argv when used from embedded python
When the library is being from embedded Python (i.e. not called directly from Python interpreter), sys.argv does not exist which makes the library crash.
To handle this rare edge case, we check that argv is there and we do the import.
If not, we simply return None.

My concrete use case is to use Google Python API with Postgres/Multicorn.